backend/requirements.txt
Flask
psycopg2-binary
google-cloud-storage